backend/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
anthropic==0.42.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
